+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- *Profile Container* fixed submit on enter 
+
 ## [2.6.3] - 2019-06-10
 
 ### Fixed